requirements.txt
# --- dependencies to be downloaded from this requirements.txt file ---

requests
tqdm
pyliftover
python-dotenv
